refactor(bunkbot): rename files to match bluebot kebab-case convention
Rename files to use consistent kebab-case naming for compound words,
matching bluebot's naming conventions.

Changes:
- Renamed bunkbot.ts to bunk-bot.ts (matches blue-bot.ts pattern)
- Renamed clearwebhooks.ts to clear-webhooks.ts (kebab-case for compound words)
- Updated import statements in:
  - src/bunkbot/src/index.ts
  - src/bunkbot/src/commands/command-registry.ts

Benefits:
- Consistent naming conventions across all bots
- Follows established kebab-case pattern for compound words
- Easier to maintain and navigate codebase

All 96 tests pass successfully after renaming.
